cogs/requirements.txt.txt
discord.py>=2.0.0
asyncio